requirements.txt
streamlit
pandas
numpy
scikit-learn
opencv-python-headless
altair
plotly
streamlit-option-menu
Pillow
matplotlib
joblib